Add COERCE primop, ast, instruction
There are already a bunch of UNBOXED-INTEGER-TO-X-FLOAT instructions, but no
ASTs or primops for them, and they don't seem to be used. So here's this
general version that I can use to interconvert floats. Technically it will work
for integers as well, as long as the backend knows what to do with an
instruction like UNBOX FIXNUM.