Fix escape analysis on certain cell usage patterns
Basically, we want to know how many cells an enter instruction uses
based on the enter instruction (and not e.g. the enclose). This is
nontrivial to do because the information isn't right there. So we
instead just count fetch instructions.

I had assumed that fetch instructions would always directly succeed
the enter instruction, due to how segregate-lexicals generates them,
but it sometimes puts create-cells in front of them instead.

This takes care of that, but long term there should be a better way
than counting fetches.