Partial inlining: Don't just add to residual function inputs as we go
Doing so causes major performance problems, as each added input
triggers many REMOVEs and PUSHNEWs for every existing input as the
:before method on (setf cleavir-ir:inputs) struggles to keep data
usage consistent.

I haven't however implemented what I think is the right way to do it.
As described in the comment, this is just doing it for variables live
where the inlining stops. This is because I don't need it for fully
inlining functions, in which case there are no residual functions
at the end anyway.