Generate AST: Rewrite lambda list generation for dynamic-env
Uses this continuation passing style so that the dynamic env
variable is hooked up correctly. I don't like it, I think it's
ugly, but I think it was already ugly so I'm not too broken up.

The only external change is you use process-lambda-list instead
of process-required, but (a) i think that's a good idea anyway,
and (b) process-required wasn't exported so it's not really
external.