Replace array-length and array-index types with something better.
Instead, we use the new types list-index, list-length, vector-index and
vector-length.  These new types are more conservative, in that they don't
relate lengths to the value of MOST-POSITIVE-FIXNUM, but derive from the
largest expressible sequence on the actual hardware.

This change means that indices are not bounded by 2^62-1, but by something
like 2^57-1.  The advantage in doing so is that not only indices and
lengths are fixnums, but also theirs sums and differences.  This speeds up
many parts of the SICL sequence module.