type-infer: Explicitly mark values as having T type on confluence.
Relying on implicit marking by removal from table causes
non-deterministic behavior as removing type constraints from the table
causes the behavior to be dependent on the order in which basic blocks
are collected. The union operation on type constraints treats a value
not having a constraint specially to make value -> Top type mappings
implicit for space efficiency, but this interacts with confluence
badly.