Speed up loading of sicl-sequence.
This commit reduces the amount of code generated when loading sicl-sequence
by roughly a factor of four.  This is achieved by two changes.  The first
change is to prune the list of vector classes for which we generate
specialized versions.  The other change is that we avoid duplicating the
body of the WITH-KEY-FUNCTION macro.  Neither change has a significant
impact on performance.